ENH: move LocalPV handling up to module utils
